Fixing failing tests due to different states of the dbs

+++ db/migrate/20131015162822_remove_unused_types.rb
@@ -1,7 +1,11 @@
 class RemoveUnusedTypes < ActiveRecord::Migration
   def up
-    drop_table :memberships
-    drop_table :groups
+    if ActiveRecord::Base.connection.tables.include?("memberships")
+      drop_table :memberships
+    end
+    if ActiveRecord::Base.connection.tables.include?("groups")
+      drop_table :groups
+    end
   end
 
   def down
